+++ BlueStarRedStar.xcodeproj/xcuserdata/spencersymington.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,12 @@
 		<key>BlueStarRedStar.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>5</integer>
+			<integer>7</integer>
+		</dict>
+		<key>TwinStarLevelEditor.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>8</integer>
 		</dict>
 	</dict>
 </dict>
