+++ Podfile
@@ -1,6 +1,6 @@
 # Uncomment the next line to define a global platform for your project
 # platform :ios, '9.0'
-
+use_modular_headers!
 target 'BlueStarRedStar' do
   # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
   use_frameworks!
@@ -12,5 +12,13 @@ target 'BlueStarRedStar' do
 
     # Pods for testing
   end
+
+
+end
+target 'TwinStarLevelEditor' do
+  use_frameworks!
+  platform :osx, '10.13'
+  inherit! :search_paths
+  pod 'RealmSwift'
 
 end
