+++ TwinStarLevelEditor/AppDelegate.swift
@@ -1,0 +1,26 @@
+//
+//  AppDelegate.swift
+//  TwinStarLevelEditor
+//
+//  Created by Spencer Symington on 2019-04-04.
+//  Copyright © 2019 Spencer Symington. All rights reserved.
+//
+
+import Cocoa
+
+@NSApplicationMain
+class AppDelegate: NSObject, NSApplicationDelegate {
+
+
+
+  func applicationDidFinishLaunching(_ aNotification: Notification) {
+    // Insert code here to initialize your application
+  }
+
+  func applicationWillTerminate(_ aNotification: Notification) {
+    // Insert code here to tear down your application
+  }
+
+
+}
+
