+++ TwinStarLevelEditor/StaticHelper.swift
@@ -1,0 +1,7 @@
+
+
+import Foundation
+
+class StaticHelper{
+  static let widthRatio : CGFloat = 1.0
+}
