+++ TwinStarLevelEditor/ViewController.swift
@@ -1,0 +1,23 @@
+
+
+import Cocoa
+import RealmSwift
+
+
+class ViewController: NSViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+
+    // Do any additional setup after loading the view.
+  }
+
+  override var representedObject: Any? {
+    didSet {
+    // Update the view, if already loaded.
+    }
+  }
+
+
+}
+
